+++ anydex/wallet/ethereum/__init__.py
@@ -1,0 +1,3 @@
+"""
+This package contains an Ethereum wallet implementation, including methods to access the database and Ethereum backend.
+"""
